+++ src/test/exception.test.ts
@@ -40,9 +40,3 @@ test.each([
     convertNumber(input, {useTraditional: true});
   }).toThrowError(errmsg);
 });
-test('Not Supported error - traditional+capital', () => {
-  const errmsg = 'Not Supported';
-  expect(() => {
-    convertNumber(1, {useTraditional: true, useCapital: true});
-  }).toThrowError(errmsg);
-});